+++ src/fhsm_crypto.c
@@ -293,7 +293,7 @@ size_t fhsm_hash_size(fhsm_hash_t h) {
         case FHSM_HASH_SHA256: case FHSM_HASH_SHA3_256: case FHSM_HASH_SHA512_256: return 32;
         case FHSM_HASH_SHA384: case FHSM_HASH_SHA3_384: return 48;
         case FHSM_HASH_SHA512: case FHSM_HASH_SHA3_512: return 64;
-        case FHSM_HASH_SHA224: case FHSM_HASH_SHA512_224: return 28;
+        case FHSM_HASH_SHA224: case FHSM_HASH_SHA512_224: case FHSM_HASH_SHA3_224: return 28;
         case FHSM_HASH_SHA1: return 20;
         case FHSM_HASH_MD5:  return 16;
         default: return 0;
@@ -305,6 +305,7 @@ static const char *hash_name(fhsm_hash_t h) {
         case FHSM_HASH_SHA256:   return "SHA2-256";
         case FHSM_HASH_SHA384:   return "SHA2-384";
         case FHSM_HASH_SHA512:   return "SHA2-512";
+        case FHSM_HASH_SHA3_224: return "SHA3-224";
         case FHSM_HASH_SHA3_256: return "SHA3-256";
         case FHSM_HASH_SHA3_384: return "SHA3-384";
         case FHSM_HASH_SHA3_512: return "SHA3-512";